refactor(worker): #115 infrastructure adapter/config 패키지 정리
