+++ keepgoing-worker/src/main/java/com/keepgoing/keepgoing/worker/infrastructure/config/redis/WorkerRedisStreamProperties.java
@@ -1,0 +1,13 @@
+package com.keepgoing.keepgoing.worker.infrastructure.config.redis;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "image-processing.streams")
+public record WorkerRedisStreamProperties(
+		boolean enabled,
+		String request,
+		String result,
+		String requestGroup,
+		String requestConsumer
+) {
+}
