+++ keepgoing-worker/src/main/java/com/keepgoing/keepgoing/worker/infrastructure/config/storage/StorageProperties.java
@@ -1,0 +1,18 @@
+package com.keepgoing.keepgoing.worker.infrastructure.config.storage;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@ConfigurationProperties(prefix = "storage")
+public record StorageProperties(
+		String endpoint,
+		String accessKey,
+		String secretKey,
+		String region,
+		BucketNames bucketNames
+) {
+	public record BucketNames(
+			String quarantine,
+			String secure
+	) {
+	}
+}
