+++ keepgoing-worker/src/main/java/com/keepgoing/keepgoing/worker/infrastructure/config/time/TimeConfig.java
@@ -1,0 +1,17 @@
+package com.keepgoing.keepgoing.worker.infrastructure.config.time;
+
+import java.time.Clock;
+import java.time.ZoneId;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class TimeConfig {
+
+	private static final ZoneId KST = ZoneId.of("Asia/Seoul");
+
+	@Bean
+	public Clock clock() {
+		return Clock.system(KST);
+	}
+}
