--- keepgoing-worker/src/main/java/com/keepgoing/keepgoing/worker/infrastructure/redis/adapter/in/ImageProcessingRequestMapper.java
@@ -1,20 +1,0 @@
-package com.keepgoing.keepgoing.worker.infrastructure.redis.adapter.in;
-
-import com.keepgoing.keepgoing.common.image.event.ImageProcessingRequestedEvent;
-import com.keepgoing.keepgoing.worker.application.port.in.ImageProcessingCommand;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-final class ImageProcessingRequestMapper {
-
-	static ImageProcessingCommand toCommand(ImageProcessingRequestedEvent event) {
-		return new ImageProcessingCommand(
-				event.publicId(),
-				event.storageKey(),
-				event.contentType(),
-				event.fileSize(),
-				event.requestedAt()
-		);
-	}
-}
