--- keepgoing-worker/src/main/java/com/keepgoing/keepgoing/worker/infrastructure/storage/StorageProperties.java
@@ -1,18 +1,0 @@
-package com.keepgoing.keepgoing.worker.infrastructure.storage;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "storage")
-public record StorageProperties(
-		String endpoint,
-		String accessKey,
-		String secretKey,
-		String region,
-		BucketNames bucketNames
-) {
-	public record BucketNames(
-			String quarantine,
-			String secure
-	) {
-	}
-}
